move stratification
